Retry machine pool scaling when there are conflicts - When scaling a machine pool there's lots of status updates to the cluster - This means hitting a 409 conflict error when trying to scale an already scaling pool is likely - There already exists a way to check for actual conflicts and retrying the save, though this was only for the CreateEditView component - That process has now become generic and is used when scaling pools - Alos fix bug where scale errors weren't caught - fixes #5029
